[cellular] Implement cellular status, missing CLI commands, and SMS PDU support
Implement cellular CLI commands and status reporting in netsim-next:
- Add `gsm sim` and `gsm tech` commands to next/cli.
- Add `sms pdu` command to next/cli to send raw hex-encoded SMS PDUs.
- Copy over cellular status fields (SMS count, RSSI, BER, registration status)
- Support `ReceivePdu` action in cell.proto, grpc-server, and cell-actor
- Add input validation for phone numbers, SMS senders, and PDU hex format to prevent silent failures.

Bug: 541261575
Test: bazel test @netsim//next/...

+++ next/modem-rs/src/modem.rs
@@ -353,8 +353,8 @@ impl ModemImpl {
         effects
     }
 
-    pub fn get_sms_count(&self) -> usize {
-        self.sim_service.get_sms_count() + self.sms_service.get_sms_count()
+    pub fn get_sms_count(&self) -> u32 {
+        (self.sim_service.get_sms_count() + self.sms_service.get_sms_count()) as u32
     }
 
     pub fn is_ringing(&self) -> bool {
